requirements/mongo-requirements.txt
pymongo
python-dotenv
py-cpuinfo